[#3125] Moving GAE specific code to its own package
