+++ GAE/src/org/waterforpeople/mapping/app/web/rest/security/google/GoogleAccountsAuthenticationProvider.java
@@ -1,5 +1,5 @@
 
-package org.waterforpeople.mapping.app.web.rest.security;
+package org.waterforpeople.mapping.app.web.rest.security.google;
 
 import java.util.EnumSet;
 import java.util.Set;
@@ -15,6 +15,8 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.SpringSecurityMessageSource;
 import org.springframework.security.web.authentication.preauth.PreAuthenticatedAuthenticationToken;
+import org.waterforpeople.mapping.app.web.rest.security.AppRole;
+import org.waterforpeople.mapping.app.web.rest.security.GaeUserAuthentication;
 import org.waterforpeople.mapping.app.web.rest.security.user.GaeUser;
 
 import com.gallatinsystems.user.dao.UserDao;
